==> notebooks/01_load_and_validate_raw_data.ipynb <==
from pathlib import Path
import requests
import zipfile
import pandas as pd
import pyarrow.parquet as pq

def download_one_file_of_raw_data(year: int) -> Path:
    """"""
    URL = f'https://cdn.buenosaires.gob.ar/datosabiertos/datasets/transporte-y-obras-publicas/bicicletas-publicas/recorridos-realizados-{year}.zip'
    response= requests.get(URL)

    
    if response.status_code == 200:
        
        path = f'../data/raw/recorridos-realizados-{year}.zip'
        open(path, "wb").write(response.content)     
        print(f'descargado año {year}')

        return path
    else:
        raise Exception(f'{URL} is not available')

import pyarrow as pa

def unzip_and_convert_csv_to_parquet(year: int) -> Path:
    nombre_archivo_zip = f"../data/raw/recorridos-realizados-{year}.zip"
        # Unzip the zip file
    with zipfile.ZipFile(nombre_archivo_zip, 'r') as archivo_zip:

        # Extract the CSV file from zip
        nombre_archivo_csv = archivo_zip.namelist()[0]  # 
        archivo_zip.extractall(f"../data/raw/")

        # Convert the DataFrame to Parquet dataframe
        df = pd.read_csv(f"../data/raw/{nombre_archivo_csv}", delimiter=',', decimal=".")

        
        nombre_archivo_parquet = f"rides_{year}.parquet"
        table = pa.Table.from_pandas(df)
        pq.write_table(table, f"../data/raw/{nombre_archivo_parquet}")

        path = f'../data/raw/rides_{year}.parquet'
    return path

download_one_file_of_raw_data(year=2022)

unzip_and_convert_csv_to_parquet(year=2022)


import pandas as pd
rides = pd.read_parquet("../data/raw/rides_2022.parquet")

rides.head(10)

rides.shape

rides.id_estacion_origen.nunique()

#We keep only the columns that interest us and we rename them
rides = rides[["fecha_origen_recorrido","id_estacion_origen"]]
## Remove the "BAEcobici" part and convert to int type
rides['id_estacion_origen'] = rides['id_estacion_origen'].str.replace('BAEcobici', '').astype(int)
## Rename columns
rides.rename(columns={"fecha_origen_recorrido":"pickup_datetime","id_estacion_origen":"pickup_location_id"},inplace=True)

rides["pickup_datetime"] = pd.to_datetime(rides["pickup_datetime"], format='%Y-%m-%d %H:%M:%S')


rides.head(10)

rides.pickup_location_id.nunique()

rides["pickup_datetime"].describe()

rides = rides[rides.pickup_datetime >= '2022-01-01']
rides = rides[rides.pickup_datetime < '2023-01-01']
rides["pickup_datetime"].describe()

rides.to_parquet("../data/transformed/validated_rides_2022.parquet")
