==> notebooks/06_baseline_model.ipynb <==
%reload_ext autoreload
%autoreload 2

import pandas as pd
from src.paths import TRANSFORMED_DATA_DIR
df = pd.read_parquet(TRANSFORMED_DATA_DIR / 'tabular_data.parquet')
df

from datetime import  datetime
from src.data_split import train_test_split

X_train, y_train, X_test, y_test = train_test_split(
    df,
    cutoff_date = datetime(2022, 6, 1, 0, 0),
    targets_columns_names = [c for c in df.columns if c.startswith('rides_next_')]
)

print(f'{X_train.shape=}')
print(f'{y_train.shape=}')
print(f'{X_test.shape=}')
print(f'{y_test.shape=}')


import numpy as np


import numpy as np

class BaselineModelPreviousHours:
    """
    Prediction = actual demand observed in the last 36 hours
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.DataFrame):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> np.array:
        """"""
        return X_test.loc[:, [f"rides_previous_{i}_hour" for i in range(1, 37)]]

model = BaselineModelPreviousHours()
predictions = model.predict(X_test)
predictions


from sklearn.metrics import mean_absolute_error
mae = mean_absolute_error(y_test, predictions)
print(f'{mae=}')

import numpy as np

class BaselineModelPreviousWeek:
    """
    Prediction = actual demand observed at t - 7 days
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.DataFrame):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> np.array:
        """"""
        return X_test.loc[:, [f"rides_previous_{7*24-1+i}_hour" for i in range(1, 37)]]
    

model = BaselineModelPreviousWeek()
predictions = model.predict(X_test)
predictions

mae = mean_absolute_error(y_test, predictions)
print(f'{mae=}')

class BaselineModelLast3Weeks:
    """
    Prediction = actual demand observed at t - 7 days, t - 14 days, t - 21 days
    """
    def fit(self, X_train: pd.DataFrame, y_train: pd.DataFrame):
        pass
    
    def predict(self, X_test: pd.DataFrame) -> pd.DataFrame:
        """"""
        X_test_1=X_test.loc[:, [f"rides_previous_{7*24-1+i}_hour" for i in range(1, 37)]] 
        X_test_2=X_test.loc[:, [f"rides_previous_{2*7*24-1+i}_hour" for i in range(1, 37)]]
        X_test_2.columns=[f"rides_previous_{7*24-1+i}_hour" for i in range(1, 37)]
        X_test_3=X_test.loc[:, [f"rides_previous_{3*7*24-1+i}_hour" for i in range(1, 37)]]
        X_test_3.columns=[f"rides_previous_{7*24-1+i}_hour" for i in range(1, 37)]

        result=0.33*(
            X_test_1 + \
            X_test_2 + \
            X_test_3)
        return  result
        

model = BaselineModelLast3Weeks()
predictions = model.predict(X_test)
predictions

mae = mean_absolute_error(y_test, predictions)
print(f'{mae=}')
